requirements.txt
pandas==2.0.2
pandas-ta
numpy==1.24.3
ccxt==4.5.43
aiogram==3.22.0
python-dotenv==1.2.1
smartmoneyconcepts==0.0.26
